Initial commit
This has the server working. It prompts for a directory path and offers to create it if it doesn't exist. Requesting /files gives an array of file names in that directory, and /files/:fileName lets you download that file.

The client still needs to be implemented but I've successfully downloaded a file through my browser.