Added client functionality
It just asks for a hostname, then prompts the user to enter the number corresponding to a file in the file list that's shown. It downloads to the Downloads directory in the user's home directory with a matching file name to the original file name.

I also refactored the prompts a bit to be a little easier to use and more clear (extra blank lines where it felt nice).